feat: complete backend command migration and bump version to 1.1.3

+++ apps/desktop/src-tauri/src/models/report_motoparts.rs
@@ -1,41 +1,41 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct DashboardStats {
     pub total_sales_today: f64,
-    pub count_sales_today: i64,
-    pub open_service_orders: i64,
-    pub active_warranties: i64,
-    pub low_stock_products: i64,
+    pub count_sales_today: i32,
+    pub open_service_orders: i32,
+    pub active_warranties: i32,
+    pub low_stock_products: i32,
     pub revenue_weekly: Vec<DailyRevenue>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct DailyRevenue {
     pub date: String,
     pub amount: f64,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct ServiceOrderStats {
-    pub total_orders: i64,
+    pub total_orders: i32,
     pub by_status: Vec<StatusCount>,
     pub revenue_labor: f64,
     pub revenue_parts: f64,
     pub average_ticket: f64,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct StatusCount {
     pub status: String,
-    pub count: i64,
+    pub count: i32,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct TopItem {
     pub id: String,
@@ -44,7 +44,7 @@ pub struct TopItem {
     pub total_value: f64,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct MotopartsReport {
     pub period_start: String,
@@ -52,6 +52,6 @@ pub struct MotopartsReport {
     pub sales_total: f64,
     pub service_labour_total: f64,
     pub parts_total: f64,
-    pub warranties_count: i64,
+    pub warranties_count: i32,
     pub warranties_cost: f64,
 }
